+++ Runtime/Colliders & Rigidbodies/FindMeshCollidersInScene.cs
@@ -6,6 +6,6 @@ public class FindMeshCollidersInScene : FindCollidersInScene
     [ContextMenu(nameof(FindAllColliders))]
     protected override void FindAllColliders()
     {
-        m_allColliders = FindObjectsOfType<MeshCollider>();
+        m_allColliders = FindObjectsByType<MeshCollider>(FindObjectsSortMode.None);
     }
 }