+++ CHANGELOG.md
@@ -4,16 +4,16 @@ All notable changes to this project will be documented in this file.
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 
-## [1.0.0] - 2024-09-13
+## [1.0.0] - 2024-11-23
 
 
 ### Added
 
--
+- Additional Unity Event
 
 ### Fixed
 
--
+- Obsolete FindObjectsByType 
 
 ### Changed
 
@@ -22,3 +22,11 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 ### Removed
 
 -
+
+
+## [1.0.0] - 2024-09-13
+
+
+### Added
+
+- Init